+++ src/app/Providers.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      // With SSR, we set staleTime to avoid refetching immediately on the client
+      staleTime: 60 * 1000,
+    },
+  },
+});
+
+export const Providers = ({
+  children,
+}: Readonly<{ children: React.ReactNode }>) => {
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+};