+++ src/app/page.tsx
@@ -1,7 +1,16 @@
+import { HomeTransactions } from "@/components/Home/HomeTransactions";
 import { TopProtocolsBarList } from "@/components/Stats/TopProtocolsBarList";
 import { TransactionRow } from "@/components/Transaction/TransactionRow";
 import { getTransactions, getTransactionsStats } from "@/db/transactions";
-import { Card, Container, Heading, Separator, Text } from "@radix-ui/themes";
+import {
+  Card,
+  Container,
+  Heading,
+  Separator,
+  Text,
+  Tooltip,
+} from "@radix-ui/themes";
+import { IconInfoCircle } from "@tabler/icons-react";
 import { Fragment, Suspense } from "react";
 
 export const dynamic = "force-dynamic";
@@ -28,16 +37,32 @@ export default async function HomePage() {
 
       <div className="mt-10 grid grid-cols-2 gap-5">
         <Card size="2">
-          <Text as="div" size="2" color="gray">
+          <Text
+            className="flex items-center gap-1"
+            as="div"
+            size="2"
+            color="gray"
+          >
             Total Transactions
+            <Tooltip content="Total number of transactions that interacted with the protocols supported by stackspulse">
+              <IconInfoCircle size={14} />
+            </Tooltip>
           </Text>
           <Text as="div" mt="2" size="5" weight="medium">
             {stats.count.toLocaleString("en-US")}
           </Text>
         </Card>
         <Card size="2">
-          <Text as="div" size="2" color="gray">
+          <Text
+            className="flex items-center gap-1"
+            as="div"
+            size="2"
+            color="gray"
+          >
             Unique addresses
+            <Tooltip content="Total number of unique addresses that interacted with the protocols supported by stackspulse">
+              <IconInfoCircle size={14} />
+            </Tooltip>
           </Text>
           <Text as="div" mt="2" size="5" weight="medium">
             {stats.uniqueSenders.toLocaleString("en-US")}
@@ -45,26 +70,11 @@ export default async function HomePage() {
         </Card>
       </div>
 
-      <Suspense>
-        <TopProtocolsBarList />
-      </Suspense>
+      <TopProtocolsBarList />
 
-      <div className="mt-10">
-        <Heading as="h2" size="3" color="gray" highContrast>
-          Transactions
-        </Heading>
-        <div className="mt-4 md:space-y-4">
-          {transactions.map((transaction) => (
-            <Fragment key={transaction.txId}>
-              <TransactionRow
-                key={transaction.txId}
-                transaction={transaction}
-              />
-              <Separator className="md:hidden" my="4" size="4" />
-            </Fragment>
-          ))}
-        </div>
-      </div>
+      <Suspense>
+        <HomeTransactions />
+      </Suspense>
     </Container>
   );
 }